NFC: trf7970a: Don't require EN2 be managed by driver
Currently, the trf7970a driver assumes that there is a GPIO
connected to the EN2 pin of the trf7970a part; however, that
may not always be the case.  For some versions of the part,
it is possible that EN2 is tied low by hardware and there is
no GPIO connected to it.  Support this configuration by having
the driver ignore EN2 when the second entry in the 'ti,enable-gpios'
DTS property is missing (that is the entry that specifies the
GPIOs for the EN and EN2 pins).

Signed-off-by: Mark Greer <mgreer@animalcreek.com>