remoteproc/pruss: add instance id to each PRUSS data
The PRUSS driver provided a struct pruss handle to client drivers
using the pruss_get() API for them to operate on a PRUSS instance.
The PRU client drivers may also want to know the exact PRUSS instance
id so that they can program/configure different PRUSS instances
differently (eg: running a Dual-EMAC PRU Ethernet usecase on one
instance, and running a PRU Ethernet Switch usecase on the other).
Add an instance id to each of the PRUSS so that the PRUSS API can
be enhanced to provide this instance id to PRU client drivers. The
values match the terminology used in each of the SoC's Technical
Reference Manuals.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Jacob Stiffler <j-stiffler@ti.com>