net: prueth: fix incorrect reporting of packet drop at egress for emac
When the xmit function runs out of buffer, the right procedure is
to stop the queue and return NETDEV_TX_BUSY so that the driver can
resume transmission when buffer becomes available. Thus fix this
by calling netif_stop_queue() when this condition occurs and return
correct error code to caller. Currently switch firmware doesn't
support tx interrupt. So do this conditionally for emac for now.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>