HACK: ARM: dts: dra7xx: program timer16 to run at 100MHz source clock
Re-program abe_dpll_ck to run at 50MHz, and select the source for
timer16 to be abe_giclk. This makes the timer run at 100MHz.

The patch should be applied under the board file which requires this
setup to be used, doing this under the generic dra7xx-clocks.dtsi is
only done for demonstration purposes.

Signed-off-by: Tero Kristo <t-kristo@ti.com>