prueth: emac: add a workaround to fix netdev watchdog issue
When pru Ethernet interface is loaded, say at 95Mbps, NETDEV WATCHDOG
timer fires with following trace:-

[ 1642.085826] WARNING: CPU: 0 PID: 4 at net/sched/sch_generic.c:316
dev_watchdog+0x258/0x25c
[ 1642.094128] NETDEV WATCHDOG: eth2 (prueth): transmit queue 0 timed out

Currently driver stops the queue when buffer overflow is detected at
the egress and wake up the queue when a tx interrupt is received from
PRU. Watchdog is fired since queue is not woken up for an extended
period of time. This is suspected to be due to tx interrupt event being
lost. For now, add a workaround to fix this by not stopping the queue
when overflow happens.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>