net: ethenet: ti: prueth: introducing ptp time stamping for prueth
This patch adds the support of IEEE-1588 ordinary clock (OC) on the
PRU-ICSS ethernet ports of TI's AM572x devices. The PRU-ICSS
includes an Industrial Ethernet Peripheral (IEP) unit. This unit
is a simple 64-bit timer (counter) whose functionality, in
conjunction with TI's PRU-ICSS industrial PRP firmware, is leveraged
in the support of IEEE-1588 PTP time stamping on the PRU-ICSS
ethernet ports.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>