prueth: prp: add emac mode support
Currently PTP support is available only to PRP firmware. So to support
SAN with PTP, the PRP firmware is used in EMAC mode which is a HACK.
This hack is expected to be supported in the driver until PTP feature
becomes available in EMAC firmware. The present implementation in
PRP firmware doesn't strip last 6 bytes to support emac mode as the
assumption is that firmware receives frames without RCT in emac mode.
Though this is true for EMAC mode, with PRP mode, firmware receives
frames from Host with RCT and for SAN destination, the RCT needs to
be stripped. Since PRP firmware currently doesn't strip RCT to a
SAN destination, this fails conformance test as SAN destination is
expected to receive frame without RCT. This behavior is expected from
a DAN-P When Node Table is implemented and is used for frame routing
to SAN. So firmware needs to do

 1. If in EMAC mode, not strip RCT
 2. If in PRP mode, strip RCT if destination is SAN
 3. if in PRP mode, don't strip RCT if destination is DAN

To support this dual mode, introduce a prp_emac_mode flag which can be
used to configure a PRP firmware to operate in EMAC mode. Since this is a
HACK, currently this configuration is done through a debugfs interface
which can be configured by user in EMAC mode before the ethernet device
is brought up. PRUETH driver sets bit0 of the buffer descriptor to
indicate mode to firmware for each frame forwarded to firmware.

bit-0 - Set - EMAC mode, Clear - PRP mode

By default PRP firmware will operate in PRP mode and user has to
do following sequence to switch mode to emac :-

1. boot up the device as a PRP Ethernet device or switch to PRP mode
   from user space.
2. Switch to EMAC mode
   echo 1 > /sys/kernel/debug/prueth-eth2/prp_emac_mode
   echo 1 > /sys/kernel/debug/prueth-eth3/prp_emac_mode
3. ifconfig eth2 <ip address> up
4. ifconfig eth3 <ip address> up

Now both eth2 and eth3 will work as an emac Ethernet device.

While doing this change, add a minor improvement in the prueth_tx_enqueue()
by defining a prueth variable instead of retrieving the same from emac
at several places in the function.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>